requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage   
flake8
python-dotenv>=0.5.1
torchvision
numpy
pandas
scikit-learn
transformers
datasets
tqdm
matplotlib
pytest
